update-update router, draw Cover from App and remove App
